pyproject.toml
[tool.poetry]
name = "crypto-bots"
version = "0.1.0"
description = "My python projects structure"
authors = ["Rami Menai <rami.menai@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
nomics-python = "^3.2.0"
ethereum-gasprice = "^1.3.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.0"
pytest = "^6.2.4"
colorlog = "^5.0.1"
taskipy = "^1.8.1"
pre-commit = "^2.14.0"
flake8 = "^3.9.2"
flake8-annotations = "^2.6.2"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-string-format = "^0.3.0"
flake8-tidy-imports = "^4.3.0"
flake8-todo = "^0.7"
pep8-naming = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python -m bots"
tests = "pytest bots/tests/"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"
